Refactor database configuration to use environment variable for DATABASE_URL and update price types in Booking and Flight services to ensure consistency with response schemas

+++ api/src/schemas/flight.py
@@ -24,7 +24,7 @@ class FlightResponse(BaseModel):
     arrival_time: datetime.datetime
     airline: str
     status: str
-    price: int  # Changed from float to int to match database model
+    price: float
 
     class Config:
         from_attributes = True
